+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+- `pubspec.lock` is now included with `--remove-cache`
+- fix method call that treated `.packages` as a folder
+
 ## 0.1.2
 
 - removal of `.pub-cache` and `.packages` has the be explicitly declared now with `--remove-cache`